update Tue 06/13/2023

+++ articles/2023-06-01-the-value-objects.assets/code-snippets/2023-06-11-the-c-lang-value-object/meters.c
@@ -1,7 +1,7 @@
 #include <stdlib.h>
 #include "meters.h"
 
-meters* meters_new(float value) {
+meters* meters_new(double value) {
     meters* new_meter = (meters*)malloc(sizeof(meters));
     if (value < 0) {
         error* err = (error*)malloc(sizeof(error));
